requirements.txt
annotated-types @ file:///home/conda/feedstock_root/build_artifacts/annotated-types_1716290248287/work
anyio @ file:///home/conda/feedstock_root/build_artifacts/anyio_1717693030552/work
arcgis==1.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
box-sdk-gen==1.1.0
boxsdk==3.11.0
Brotli @ file:///D:/bld/brotli-split_1695989908365/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1720457958366/work/certifi
cffi @ file:///D:/bld/cffi_1696001932509/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
click @ file:///D:/bld/click_1692312014553/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
dnspython @ file:///home/conda/feedstock_root/build_artifacts/dnspython_1709190728464/work
email_validator @ file:///home/conda/feedstock_root/build_artifacts/email-validator-meta_1718984626658/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1720869315914/work
executing==2.0.1
fastapi @ file:///home/conda/feedstock_root/build_artifacts/fastapi_1720998438716/work
fastapi-cli @ file:///home/conda/feedstock_root/build_artifacts/fastapi-cli_1716205772095/work
fastjsonschema==2.20.0
fonttools==4.53.1
fqdn==1.5.1
h11 @ file:///home/conda/feedstock_root/build_artifacts/h11_1664132893548/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1634280454336/work
hpack==4.0.0
httpcore @ file:///home/conda/feedstock_root/build_artifacts/httpcore_1711596990900/work
httptools==0.6.1
httpx @ file:///home/conda/feedstock_root/build_artifacts/httpx_1708530890843/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1619110129307/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1713279365350/work
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1715127149914/work
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
keyring==25.2.1
kiwisolver==1.4.5
lerc==4.0.1
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1686175045316/work
MarkupSafe @ file:///D:/bld/markupsafe_1706900055959/work
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1704317613764/work
mistune==3.0.2
more-itertools==10.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook_shim==0.2.4
numpy==2.0.1
oauthlib==3.2.2
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pillow @ file:///D:/bld/pillow_1719903666313/work
platformdirs==4.2.2
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1711811537435/work
pydantic @ file:///home/conda/feedstock_root/build_artifacts/pydantic_1720293063581/work
pydantic-extra-types==2.9.0
pydantic-settings==2.3.4
pydantic_core @ file:///D:/bld/pydantic-core_1720041343998/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pyparsing==3.1.2
pyshp==2.3.1
PySocks @ file:///D:/bld/pysocks_1661604991356/work
pyspnego==0.11.1
pytest==8.3.1
python-dateutil==2.9.0.post0
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/python-dotenv-split_1706018097647/work
python-json-logger==2.0.7
python-multipart @ file:///home/conda/feedstock_root/build_artifacts/python-multipart_1707760088566/work
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1717057054362/work
requests-kerberos==0.15.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requests_ntlm==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich @ file:///home/conda/feedstock_root/build_artifacts/rich-split_1709150387247/work/dist
rpds-py==0.19.0
Send2Trash==1.8.3
setuptools==69.5.1
shellingham @ file:///home/conda/feedstock_root/build_artifacts/shellingham_1698144360966/work
six==1.16.0
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1708952932303/work
soupsieve==2.5
sspilib==0.1.0
stack-data==0.6.3
starlette @ file:///home/conda/feedstock_root/build_artifacts/starlette-recipe_1709667058396/work
terminado==0.18.1
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
typer==0.12.3
typer-slim==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.1
ujson==5.10.0
uri-template==1.3.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1719391292974/work
uvicorn==0.30.3
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
wheel==0.43.0
widgetsnbextension==4.0.11
win-inet-pton @ file:///D:/bld/win_inet_pton_1667051142467/work
winkerberos==0.11.0
zstandard==0.23.0
